Rename 'functions' directory to 'internal'
